// frontend/src/types/PetTypes.ts
export type lostPetProps = {
  _id: string;
  name: string;
  type: string;
  owner: string;
  date: string;
  location: string;
  description: string;
  image: string;
};

export type adoptPetProps = {
  _id: string;
  name: string;
  type: string;
  owner: string;
  location: string;
  description: string;
  image: string;
};